Add RPC integration tests. Split CI runners and tests (#885)
* Test for RPC interface

* Use ThreadNode with RPC

* Separate integration tests

* Configure CI to run integration tests separately

* Tag runners

* Fix CI. Fix RPC tests

* Make expensive integration tests run on schedule only

* Run unit tests and inexpensive integration tests together
