+++ Cargo.toml
@@ -68,3 +68,7 @@ codegen-units = 1 # Use only 1 codegen-unit to enable full optimizations.
 
 [profile.dev.overrides.pairing]
 opt-level = 3 # pairing library is too slow to use in debug
+
+[features]
+expensive_tests = []
+regression_tests = []
