+++ test-utils/testlib/src/lib.rs
@@ -5,3 +5,4 @@ pub mod test_helpers;
 pub mod transactions_generator;
 pub mod transactions_executor;
 pub mod user;
+pub mod standard_test_cases;
