+++ test-utils/testlib/src/standard_test_cases.rs
@@ -1,14 +1,9 @@
-use configs::chain_spec::{
-    AuthorityRotation, ChainSpec, DefaultIdType, TESTING_INIT_BALANCE, TESTING_INIT_STAKE,
-};
-use configs::ClientConfig;
-use network::proxy::benchmark::BenchmarkHandler;
-use network::proxy::ProxyHandler;
+use crate::node::{Node, RuntimeNode};
+use crate::test_helpers::wait;
+use crate::user::User;
+use configs::chain_spec::{TESTING_INIT_BALANCE, TESTING_INIT_STAKE};
 use node_runtime::state_viewer::AccountViewCallResult;
-use node_runtime::test_utils::{
-    alice_account, bob_account, default_code_hash, encode_int, eve_account,
-};
-use storage::set;
+use node_runtime::test_utils::{bob_account, default_code_hash, encode_int, eve_account};
 use primitives::crypto::signer::InMemorySigner;
 use primitives::hash::{hash, CryptoHash};
 use primitives::serialize::Decode;
@@ -19,75 +14,12 @@ use primitives::transaction::{
     TransactionStatus,
 };
 use primitives::types::AccountingInfo;
-use std::sync::atomic::{AtomicU16, Ordering};
-use std::sync::Arc;
-use testlib::node::{
-    create_nodes_with_id_type, Node, NodeConfig, RuntimeNode, ShardClientNode, ThreadNode,
-    TEST_BLOCK_FETCH_LIMIT, TEST_BLOCK_MAX_SIZE,
-};
-use testlib::test_helpers::{heavy_test, wait};
-use testlib::user::User;
 use primitives::utils::key_for_callback;
-
-const POISONED_LOCK_ERR: &str = "The lock was poisoned.";
-const NUM_TEST_NODE: usize = 4;
-static TEST_PORT: AtomicU16 = AtomicU16::new(6000);
-
-fn test_chain_spec() -> ChainSpec {
-    ChainSpec::testing_spec(DefaultIdType::Named, 3, 3, AuthorityRotation::ProofOfAuthority).0
-}
-
-fn create_shard_client_node() -> ShardClientNode {
-    let mut client_cfg = ClientConfig::default_devnet();
-    client_cfg.chain_spec = test_chain_spec();
-    ShardClientNode::new(client_cfg)
-}
-
-fn create_runtime_node() -> RuntimeNode {
-    RuntimeNode::new(&alice_account())
-}
-
-fn create_testnet_node(test_prefix: &str, test_port: u16) -> Vec<ThreadNode> {
-    let proxy_handlers: Vec<Arc<ProxyHandler>> = vec![Arc::new(BenchmarkHandler::new())];
-
-    let (_, account_names, mut nodes) = create_nodes_with_id_type(
-        NUM_TEST_NODE,
-        test_prefix,
-        test_port,
-        TEST_BLOCK_FETCH_LIMIT,
-        TEST_BLOCK_MAX_SIZE,
-        proxy_handlers,
-        DefaultIdType::Named,
-    );
-    assert_eq!(account_names[0], alice_account());
-    let mut nodes: Vec<_> = nodes
-        .into_iter()
-        .map(|cfg| match cfg {
-            NodeConfig::Thread(config) => ThreadNode::new(config),
-            _ => unreachable!(),
-        })
-        .collect();
-    for i in 0..NUM_TEST_NODE {
-        nodes[i].start();
-    }
-    nodes
-}
-
-/// Macro for running testnet test. Increment the atomic global counter for port,
-/// and get the test_prefix from the test name.
-macro_rules! run_testnet_test {
-    ($f:expr) => {
-        let port = TEST_PORT.fetch_add(NUM_TEST_NODE as u16, Ordering::SeqCst);
-        let test_prefix = stringify!($f);
-        let mut nodes = create_testnet_node(test_prefix, port);
-        let node = nodes.pop().unwrap();
-        heavy_test(|| $f(node));
-    };
-}
+use storage::set;
 
 /// validate transaction result in the case that it is successfully and generate one receipt which
-/// itself generates another receipt.
-fn validate_tx_result(node_user: Box<User>, root: CryptoHash, hash: &CryptoHash) {
+/// itself generates another receipt. sfdsa
+pub fn validate_tx_result(node_user: Box<User>, root: CryptoHash, hash: &CryptoHash) {
     let transaction_result = node_user.get_transaction_result(&hash);
     assert_eq!(transaction_result.status, TransactionStatus::Completed);
     assert_eq!(transaction_result.receipts.len(), 1);
@@ -102,7 +34,7 @@ fn validate_tx_result(node_user: Box<User>, root: CryptoHash, hash: &CryptoHash)
 }
 
 /// Wait until transaction finishes (either succeeds or fails).
-fn wait_for_transaction(node_user: &Box<User>, hash: &CryptoHash) {
+pub fn wait_for_transaction(node_user: &Box<User>, hash: &CryptoHash) {
     wait(
         || match node_user.get_transaction_final_result(hash).status {
             FinalTransactionStatus::Unknown | FinalTransactionStatus::Started => false,
@@ -113,7 +45,7 @@ fn wait_for_transaction(node_user: &Box<User>, hash: &CryptoHash) {
     );
 }
 
-fn test_smart_contract_simple(node: impl Node) {
+pub fn test_smart_contract_simple(node: impl Node) {
     let account_id = &node.signer().account_id;
     let transaction = TransactionBody::FunctionCall(FunctionCallTransaction {
         nonce: node.get_account_nonce(account_id).unwrap_or_default() + 1,
@@ -133,7 +65,7 @@ fn test_smart_contract_simple(node: impl Node) {
     validate_tx_result(node_user, root, &hash);
 }
 
-fn test_smart_contract_bad_method_name(node: impl Node) {
+pub fn test_smart_contract_bad_method_name(node: impl Node) {
     let account_id = &node.signer().account_id;
     let transaction = TransactionBody::FunctionCall(FunctionCallTransaction {
         nonce: node.get_account_nonce(account_id).unwrap_or_default() + 1,
@@ -158,7 +90,7 @@ fn test_smart_contract_bad_method_name(node: impl Node) {
     assert_ne!(root, new_root);
 }
 
-fn test_smart_contract_empty_method_name_with_no_tokens(node: impl Node) {
+pub fn test_smart_contract_empty_method_name_with_no_tokens(node: impl Node) {
     let account_id = &node.signer().account_id;
     let transaction = TransactionBody::FunctionCall(FunctionCallTransaction {
         nonce: node.get_account_nonce(account_id).unwrap_or_default() + 1,
@@ -183,7 +115,7 @@ fn test_smart_contract_empty_method_name_with_no_tokens(node: impl Node) {
     assert_ne!(root, new_root);
 }
 
-fn test_smart_contract_empty_method_name_with_tokens(node: impl Node) {
+pub fn test_smart_contract_empty_method_name_with_tokens(node: impl Node) {
     let account_id = &node.signer().account_id;
     let transaction = TransactionBody::FunctionCall(FunctionCallTransaction {
         nonce: node.get_account_nonce(account_id).unwrap_or_default() + 1,
@@ -203,7 +135,7 @@ fn test_smart_contract_empty_method_name_with_tokens(node: impl Node) {
     validate_tx_result(node_user, root, &hash);
 }
 
-fn test_smart_contract_with_args(node: impl Node) {
+pub fn test_smart_contract_with_args(node: impl Node) {
     let account_id = &node.signer().account_id;
     let transaction = TransactionBody::FunctionCall(FunctionCallTransaction {
         nonce: node.get_account_nonce(account_id).unwrap_or_default() + 1,
@@ -223,7 +155,7 @@ fn test_smart_contract_with_args(node: impl Node) {
     validate_tx_result(node_user, root, &hash);
 }
 
-fn test_async_call_with_no_callback(node: impl Node) {
+pub fn test_async_call_with_no_callback(node: impl Node) {
     let account_id = &node.signer().account_id;
     let nonce = hash(&[1, 2, 3]);
     let receipt = ReceiptTransaction {
@@ -255,7 +187,7 @@ fn test_async_call_with_no_callback(node: impl Node) {
     assert_ne!(root, new_root);
 }
 
-fn test_async_call_with_callback(node: impl Node) {
+pub fn test_async_call_with_callback(node: impl Node) {
     let account_id = &node.signer().account_id;
     let args = (7..9).flat_map(|x| encode_int(x).to_vec()).collect();
     let accounting_info = AccountingInfo { originator: account_id.clone(), contract_id: None };
@@ -304,7 +236,7 @@ fn test_async_call_with_callback(node: impl Node) {
     }
 }
 
-fn test_async_call_with_logs(node: impl Node) {
+pub fn test_async_call_with_logs(node: impl Node) {
     let account_id = &node.signer().account_id;
     let nonce = hash(&[1, 2, 3]);
     let receipt = ReceiptTransaction {
@@ -337,7 +269,7 @@ fn test_async_call_with_logs(node: impl Node) {
     assert_ne!(root, new_root);
 }
 
-fn test_deposit_with_callback(node: impl Node) {
+pub fn test_deposit_with_callback(node: impl Node) {
     let account_id = &node.signer().account_id;
     let args = (7..9).flat_map(|x| encode_int(x).to_vec()).collect();
     let accounting_info = AccountingInfo { originator: account_id.clone(), contract_id: None };
@@ -370,7 +302,7 @@ fn test_deposit_with_callback(node: impl Node) {
 }
 
 // This test only works with RuntimeNode because it requires modifying state.
-fn test_callback(node: RuntimeNode) {
+pub fn test_callback(node: RuntimeNode) {
     let account_id = &node.signer().account_id;
     let accounting_info = AccountingInfo { originator: account_id.clone(), contract_id: None };
     let mut callback =
@@ -378,11 +310,11 @@ fn test_callback(node: RuntimeNode) {
     callback.results.resize(1, None);
     let callback_id = [0; 32].to_vec();
 
-    let mut state_update = node.client.read().expect(POISONED_LOCK_ERR).get_state_update();
+    let mut state_update = node.client.read().unwrap().get_state_update();
     set(&mut state_update, &key_for_callback(&callback_id), &callback);
     let (root, transaction) = state_update.finalize();
     {
-        let mut client = node.client.write().expect(POISONED_LOCK_ERR);
+        let mut client = node.client.write().unwrap();
         client.state_root = root;
         client.trie.apply_changes(transaction).unwrap();
     }
@@ -413,7 +345,7 @@ fn test_callback(node: RuntimeNode) {
     assert_ne!(root, new_root);
 }
 
-fn test_callback_failure(node: RuntimeNode) {
+pub fn test_callback_failure(node: RuntimeNode) {
     let account_id = &node.signer().account_id;
     let accounting_info = AccountingInfo { originator: account_id.clone(), contract_id: None };
     let mut callback = Callback::new(
@@ -424,11 +356,11 @@ fn test_callback_failure(node: RuntimeNode) {
     );
     callback.results.resize(1, None);
     let callback_id = [0; 32].to_vec();
-    let mut state_update = node.client.read().expect(POISONED_LOCK_ERR).get_state_update();
+    let mut state_update = node.client.read().unwrap().get_state_update();
     set(&mut state_update, &key_for_callback(&callback_id.clone()), &callback);
     let (root, transaction) = state_update.finalize();
     {
-        let mut client = node.client.write().expect(POISONED_LOCK_ERR);
+        let mut client = node.client.write().unwrap();
         client.state_root = root;
         client.trie.apply_changes(transaction).unwrap();
     }
@@ -459,9 +391,9 @@ fn test_callback_failure(node: RuntimeNode) {
     assert_ne!(root, new_root);
 }
 
-fn test_nonce_update_when_deploying_contract(node: impl Node) {
+pub fn test_nonce_update_when_deploying_contract(node: impl Node) {
     let account_id = &node.signer().account_id;
-    let wasm_binary = include_bytes!("../core/wasm/runtest/res/wasm_with_mem.wasm");
+    let wasm_binary = include_bytes!("../../../core/wasm/runtest/res/wasm_with_mem.wasm");
     let transaction = TransactionBody::DeployContract(DeployContractTransaction {
         nonce: node.get_account_nonce(account_id).unwrap_or_default() + 1,
         contract_id: account_id.clone(),
@@ -482,7 +414,7 @@ fn test_nonce_update_when_deploying_contract(node: impl Node) {
     assert_ne!(root, new_root);
 }
 
-fn test_nonce_updated_when_tx_failed(node: impl Node) {
+pub fn test_nonce_updated_when_tx_failed(node: impl Node) {
     let account_id = &node.signer().account_id;
     let transaction = TransactionBody::send_money(
         node.get_account_nonce(account_id).unwrap_or_default() + 1,
@@ -505,7 +437,7 @@ fn test_nonce_updated_when_tx_failed(node: impl Node) {
     assert_ne!(root, new_root);
 }
 
-fn test_upload_contract(node: impl Node) {
+pub fn test_upload_contract(node: impl Node) {
     let account_id = &node.signer().account_id;
     let node_user = node.user();
     let root = node_user.get_state_root();
@@ -523,7 +455,7 @@ fn test_upload_contract(node: impl Node) {
 
     let new_root = node_user.get_state_root();
     assert_ne!(root, new_root);
-    let wasm_binary = include_bytes!("../core/wasm/runtest/res/wasm_with_mem.wasm");
+    let wasm_binary = include_bytes!("../../../core/wasm/runtest/res/wasm_with_mem.wasm");
     let transaction = TransactionBody::DeployContract(DeployContractTransaction {
         nonce: 1,
         contract_id: eve_account(),
@@ -545,7 +477,7 @@ fn test_upload_contract(node: impl Node) {
     assert_eq!(account.code_hash, hash(wasm_binary));
 }
 
-fn test_redeploy_contract(node: impl Node) {
+pub fn test_redeploy_contract(node: impl Node) {
     let account_id = &node.signer().account_id;
     let node_user = node.user();
     let test_binary = b"test_binary";
@@ -569,7 +501,7 @@ fn test_redeploy_contract(node: impl Node) {
     assert_eq!(account.code_hash, hash(test_binary));
 }
 
-fn test_send_money(node: impl Node) {
+pub fn test_send_money(node: impl Node) {
     let account_id = &node.signer().account_id;
     let node_user = node.user();
     let money_used = 10;
@@ -617,7 +549,7 @@ fn test_send_money(node: impl Node) {
     );
 }
 
-fn test_send_money_over_balance(node: impl Node) {
+pub fn test_send_money_over_balance(node: impl Node) {
     let account_id = &node.signer().account_id;
     let node_user = node.user();
     let money_used = TESTING_INIT_BALANCE + 1;
@@ -666,7 +598,7 @@ fn test_send_money_over_balance(node: impl Node) {
     );
 }
 
-fn test_refund_on_send_money_to_non_existent_account(node: impl Node) {
+pub fn test_refund_on_send_money_to_non_existent_account(node: impl Node) {
     let account_id = &node.signer().account_id;
     let node_user = node.user();
     let money_used = 10;
@@ -711,7 +643,7 @@ fn test_refund_on_send_money_to_non_existent_account(node: impl Node) {
     assert!(result2.is_err());
 }
 
-fn test_create_account(node: impl Node) {
+pub fn test_create_account(node: impl Node) {
     let account_id = &node.signer().account_id;
     let node_user = node.user();
     let root = node_user.get_state_root();
@@ -765,7 +697,7 @@ fn test_create_account(node: impl Node) {
     );
 }
 
-fn test_create_account_again(node: impl Node) {
+pub fn test_create_account_again(node: impl Node) {
     let account_id = &node.signer().account_id;
     let node_user = node.user();
     let money_used = 10;
@@ -848,7 +780,7 @@ fn test_create_account_again(node: impl Node) {
     );
 }
 
-fn test_create_account_failure_invalid_name(node: impl Node) {
+pub fn test_create_account_failure_invalid_name(node: impl Node) {
     let account_id = &node.signer().account_id;
     let node_user = node.user();
     let mut root = node_user.get_state_root();
@@ -894,7 +826,7 @@ fn test_create_account_failure_invalid_name(node: impl Node) {
     }
 }
 
-fn test_create_account_failure_already_exists(node: impl Node) {
+pub fn test_create_account_failure_already_exists(node: impl Node) {
     let account_id = &node.signer().account_id;
     let node_user = node.user();
     let root = node_user.get_state_root();
@@ -952,7 +884,7 @@ fn test_create_account_failure_already_exists(node: impl Node) {
     );
 }
 
-fn test_swap_key(node: impl Node) {
+pub fn test_swap_key(node: impl Node) {
     let account_id = &node.signer().account_id;
     let signer2 = InMemorySigner::from_random();
     let node_user = node.user();
@@ -993,7 +925,7 @@ fn test_swap_key(node: impl Node) {
     assert_eq!(account.public_keys, vec![signer2.public_key]);
 }
 
-fn test_add_key(node: impl Node) {
+pub fn test_add_key(node: impl Node) {
     let account_id = &node.signer().account_id;
     let signer2 = InMemorySigner::from_random();
     let node_user = node.user();
@@ -1019,7 +951,7 @@ fn test_add_key(node: impl Node) {
     assert_eq!(account.public_keys[1].clone(), signer2.public_key);
 }
 
-fn test_add_existing_key(node: impl Node) {
+pub fn test_add_existing_key(node: impl Node) {
     let account_id = &node.signer().account_id;
     let node_user = node.user();
     let root = node_user.get_state_root();
@@ -1043,7 +975,7 @@ fn test_add_existing_key(node: impl Node) {
     assert_eq!(account.public_keys.len(), 1);
 }
 
-fn test_delete_key(node: impl Node) {
+pub fn test_delete_key(node: impl Node) {
     let account_id = &node.signer().account_id;
     let signer2 = InMemorySigner::from_random();
     let node_user = node.user();
@@ -1085,7 +1017,7 @@ fn test_delete_key(node: impl Node) {
     assert_eq!(account.public_keys[0].clone(), signer2.public_key);
 }
 
-fn test_delete_key_not_owned(node: impl Node) {
+pub fn test_delete_key_not_owned(node: impl Node) {
     let account_id = &node.signer().account_id;
     let signer2 = InMemorySigner::from_random();
     let node_user = node.user();
@@ -1111,7 +1043,7 @@ fn test_delete_key_not_owned(node: impl Node) {
     assert_eq!(account.public_keys.len(), 1);
 }
 
-fn test_delete_key_no_key_left(node: impl Node) {
+pub fn test_delete_key_no_key_left(node: impl Node) {
     let account_id = &node.signer().account_id;
     let node_user = node.user();
     let root = node_user.get_state_root();
@@ -1134,463 +1066,4 @@ fn test_delete_key_no_key_left(node: impl Node) {
 
     let account = node_user.view_account(account_id).unwrap();
     assert_eq!(account.public_keys.len(), 1);
-}
-
-#[cfg(test)]
-mod test {
-    use super::*;
-
-    #[test]
-    fn test_smart_contract_simple_runtime() {
-        let node = create_runtime_node();
-        test_smart_contract_simple(node);
-    }
-
-    #[test]
-    fn test_smart_contract_simple_shard_client() {
-        let node = create_shard_client_node();
-        test_smart_contract_simple(node);
-    }
-
-    #[test]
-    fn test_smart_contract_simple_testnet() {
-        run_testnet_test!(test_smart_contract_simple);
-    }
-
-    #[test]
-    fn test_smart_contract_bad_method_name_runtime() {
-        let node = create_runtime_node();
-        test_smart_contract_bad_method_name(node);
-    }
-
-    #[test]
-    fn test_smart_contract_bad_method_name_shard_client() {
-        let node = create_shard_client_node();
-        test_smart_contract_bad_method_name(node);
-    }
-
-    #[test]
-    fn test_smart_contract_bad_method_name_testnet() {
-        run_testnet_test!(test_smart_contract_bad_method_name);
-    }
-
-    #[test]
-    fn test_smart_contract_empty_method_name_with_no_tokens_runtime() {
-        let node = create_runtime_node();
-        test_smart_contract_empty_method_name_with_no_tokens(node);
-    }
-
-    #[test]
-    fn test_smart_contract_empty_method_name_with_no_tokens_shard_client() {
-        let node = create_shard_client_node();
-        test_smart_contract_empty_method_name_with_no_tokens(node);
-    }
-
-    #[test]
-    fn test_smart_contract_empty_method_name_with_no_tokens_testnet() {
-        run_testnet_test!(test_smart_contract_empty_method_name_with_no_tokens);
-    }
-
-    #[test]
-    fn test_smart_contract_empty_method_name_with_tokens_runtime() {
-        let node = create_runtime_node();
-        test_smart_contract_empty_method_name_with_tokens(node);
-    }
-
-    #[test]
-    fn test_smart_contract_empty_method_name_with_tokens_shard_client() {
-        let node = create_shard_client_node();
-        test_smart_contract_empty_method_name_with_tokens(node);
-    }
-
-    #[test]
-    fn test_smart_contract_empty_method_name_with_tokens_testnet() {
-        run_testnet_test!(test_smart_contract_empty_method_name_with_tokens);
-    }
-
-    #[test]
-    fn test_smart_contract_with_args_runtime() {
-        let node = create_runtime_node();
-        test_smart_contract_with_args(node);
-    }
-
-    #[test]
-    fn test_smart_contract_with_args_shard_client() {
-        let node = create_shard_client_node();
-        test_smart_contract_with_args(node);
-    }
-
-    #[test]
-    fn test_smart_contract_with_args_testnet() {
-        run_testnet_test!(test_smart_contract_with_args);
-    }
-
-    #[test]
-    fn test_async_call_with_no_callback_runtime() {
-        let node = create_runtime_node();
-        test_async_call_with_no_callback(node);
-    }
-
-    #[test]
-    fn test_async_call_with_no_callback_shard_client() {
-        let node = create_shard_client_node();
-        test_async_call_with_no_callback(node);
-    }
-
-    #[test]
-    fn test_async_call_with_no_callback_testnet() {
-        run_testnet_test!(test_async_call_with_no_callback);
-    }
-
-    #[test]
-    fn test_async_call_with_callback_runtime() {
-        let node = create_runtime_node();
-        test_async_call_with_callback(node);
-    }
-
-    #[test]
-    fn test_async_call_with_callback_shard_client() {
-        let node = create_shard_client_node();
-        test_async_call_with_callback(node);
-    }
-
-    #[test]
-    fn test_async_call_with_callback_testnet() {
-        run_testnet_test!(test_async_call_with_callback);
-    }
-
-    #[test]
-    fn test_async_call_with_logs_runtime() {
-        let node = create_runtime_node();
-        test_async_call_with_logs(node);
-    }
-
-    #[test]
-    fn test_async_call_with_logs_shard_client() {
-        let node = create_shard_client_node();
-        test_async_call_with_logs(node);
-    }
-
-    #[test]
-    fn test_async_call_with_logs_testnet() {
-        run_testnet_test!(test_async_call_with_logs);
-    }
-
-    #[test]
-    fn test_callback_runtime() {
-        let node = create_runtime_node();
-        test_callback(node);
-    }
-
-    #[test]
-    fn test_callback_failure_runtime() {
-        let node = create_runtime_node();
-        test_callback_failure(node);
-    }
-
-    #[test]
-    fn test_deposit_with_callback_runtime() {
-        let node = create_runtime_node();
-        test_deposit_with_callback(node);
-    }
-
-    #[test]
-    fn test_deposit_with_callback_shard_client() {
-        let node = create_shard_client_node();
-        test_deposit_with_callback(node);
-    }
-
-    #[test]
-    fn test_deposit_with_callback_testnet() {
-        run_testnet_test!(test_deposit_with_callback);
-    }
-
-    #[test]
-    fn test_nonce_update_when_deploying_contract_runtime() {
-        let node = create_runtime_node();
-        test_nonce_update_when_deploying_contract(node);
-    }
-
-    #[test]
-    fn test_nonce_update_when_deploying_contract_shard_client() {
-        let node = create_shard_client_node();
-        test_nonce_update_when_deploying_contract(node);
-    }
-
-    #[test]
-    fn test_nonce_update_when_deploying_contract_testnet() {
-        run_testnet_test!(test_nonce_update_when_deploying_contract);
-    }
-
-    #[test]
-    fn test_nonce_updated_when_tx_failed_runtime() {
-        let node = create_runtime_node();
-        test_nonce_updated_when_tx_failed(node);
-    }
-
-    #[test]
-    fn test_nonce_updated_when_tx_failed_shard_client() {
-        let node = create_shard_client_node();
-        test_nonce_updated_when_tx_failed(node);
-    }
-
-    #[test]
-    fn test_nonce_updated_when_tx_failed_testnet() {
-        run_testnet_test!(test_nonce_updated_when_tx_failed);
-    }
-
-    #[test]
-    fn test_upload_contract_runtime() {
-        let node = create_runtime_node();
-        test_upload_contract(node);
-    }
-
-    #[test]
-    fn test_upload_contract_shard_client() {
-        let node = create_shard_client_node();
-        test_upload_contract(node);
-    }
-
-    #[test]
-    fn test_upload_contract_testnet() {
-        run_testnet_test!(test_upload_contract);
-    }
-
-    #[test]
-    fn test_redeploy_contract_runtime() {
-        let node = create_runtime_node();
-        test_redeploy_contract(node);
-    }
-
-    #[test]
-    fn test_redeploy_contract_shard_client() {
-        let node = create_shard_client_node();
-        test_redeploy_contract(node);
-    }
-
-    #[test]
-    fn test_redeploy_contract_testnet() {
-        run_testnet_test!(test_redeploy_contract);
-    }
-
-    #[test]
-    fn test_send_money_runtime() {
-        let node = create_runtime_node();
-        test_send_money(node);
-    }
-
-    #[test]
-    fn test_send_money_shard_client() {
-        let node = create_shard_client_node();
-        test_send_money(node);
-    }
-
-    #[test]
-    fn test_send_money_testnet() {
-        run_testnet_test!(test_send_money);
-    }
-
-    #[test]
-    fn test_send_money_over_balance_runtime() {
-        let node = create_runtime_node();
-        test_send_money_over_balance(node);
-    }
-
-    #[test]
-    fn test_send_money_over_balance_shard_client() {
-        let node = create_shard_client_node();
-        test_send_money_over_balance(node);
-    }
-
-    #[test]
-    fn test_send_money_over_balance_testnet() {
-        run_testnet_test!(test_send_money_over_balance);
-    }
-
-    #[test]
-    fn test_refund_on_send_money_to_non_existent_account_runtime() {
-        let node = create_runtime_node();
-        test_refund_on_send_money_to_non_existent_account(node);
-    }
-
-    #[test]
-    fn test_refund_on_send_money_to_non_existent_account_shard_client() {
-        let node = create_shard_client_node();
-        test_refund_on_send_money_to_non_existent_account(node);
-    }
-
-    #[test]
-    fn test_refund_on_send_money_to_non_existent_account_testnet() {
-        run_testnet_test!(test_refund_on_send_money_to_non_existent_account);
-    }
-
-    #[test]
-    fn test_create_account_runtime() {
-        let node = create_runtime_node();
-        test_create_account(node);
-    }
-
-    #[test]
-    fn test_create_account_shard_client() {
-        let node = create_shard_client_node();
-        test_create_account(node);
-    }
-
-    #[test]
-    fn test_create_account_testnet() {
-        run_testnet_test!(test_create_account);
-    }
-
-    #[test]
-    fn test_create_account_again_runtime() {
-        let node = create_runtime_node();
-        test_create_account_again(node);
-    }
-
-    #[test]
-    fn test_create_account_again_shard_client() {
-        let node = create_shard_client_node();
-        test_create_account_again(node);
-    }
-
-    #[test]
-    fn test_create_account_again_testnet() {
-        run_testnet_test!(test_create_account_again);
-    }
-
-    #[test]
-    fn test_create_account_failure_invalid_name_runtime() {
-        let node = create_runtime_node();
-        test_create_account_failure_invalid_name(node);
-    }
-
-    #[test]
-    fn test_create_account_failure_invalid_name_shard_client() {
-        let node = create_shard_client_node();
-        test_create_account_failure_invalid_name(node);
-    }
-
-    #[test]
-    fn test_create_account_failure_invalid_name_testnet() {
-        run_testnet_test!(test_create_account_failure_invalid_name);
-    }
-
-    #[test]
-    fn test_create_account_failure_already_exists_runtime() {
-        let node = create_runtime_node();
-        test_create_account_failure_already_exists(node);
-    }
-
-    #[test]
-    fn test_create_account_failure_already_exists_shard_client() {
-        let node = create_shard_client_node();
-        test_create_account_failure_already_exists(node);
-    }
-
-    #[test]
-    fn test_create_account_failure_already_exists_testnet() {
-        run_testnet_test!(test_create_account_failure_already_exists);
-    }
-
-    #[test]
-    fn test_swap_key_runtime() {
-        let node = create_runtime_node();
-        test_swap_key(node);
-    }
-
-    #[test]
-    fn test_swap_key_shard_client() {
-        let node = create_shard_client_node();
-        test_swap_key(node);
-    }
-
-    #[test]
-    fn test_swap_key_testnet() {
-        run_testnet_test!(test_swap_key);
-    }
-
-    #[test]
-    fn test_add_key_runtime() {
-        let node = create_runtime_node();
-        test_add_key(node);
-    }
-
-    #[test]
-    fn test_add_key_shard_client() {
-        let node = create_shard_client_node();
-        test_add_key(node);
-    }
-
-    #[test]
-    fn test_add_key_testnet() {
-        run_testnet_test!(test_add_key);
-    }
-
-    #[test]
-    fn test_add_existing_key_runtime() {
-        let node = create_runtime_node();
-        test_add_existing_key(node);
-    }
-
-    #[test]
-    fn test_add_existing_key_shard_client() {
-        let node = create_shard_client_node();
-        test_add_existing_key(node);
-    }
-
-    #[test]
-    fn test_add_existing_key_testnet() {
-        run_testnet_test!(test_add_existing_key);
-    }
-
-    #[test]
-    fn test_delete_key_runtime() {
-        let node = create_runtime_node();
-        test_delete_key(node);
-    }
-
-    #[test]
-    fn test_delete_key_shard_client() {
-        let node = create_shard_client_node();
-        test_delete_key(node);
-    }
-
-    #[test]
-    fn test_delete_key_testnet() {
-        run_testnet_test!(test_delete_key);
-    }
-
-    #[test]
-    fn test_delete_key_not_owned_runtime() {
-        let node = create_runtime_node();
-        test_delete_key_not_owned(node);
-    }
-
-    #[test]
-    fn test_delete_key_not_owned_shard_client() {
-        let node = create_shard_client_node();
-        test_delete_key_not_owned(node);
-    }
-
-    #[test]
-    fn test_delete_key_not_owned_testnet() {
-        run_testnet_test!(test_delete_key_not_owned);
-    }
-
-    #[test]
-    fn test_delete_key_no_key_left_runtime() {
-        let node = create_runtime_node();
-        test_delete_key_no_key_left(node);
-    }
-
-    #[test]
-    fn test_delete_key_no_key_left_shard_client() {
-        let node = create_shard_client_node();
-        test_delete_key_no_key_left(node);
-    }
-
-    #[test]
-    fn test_delete_key_no_key_left_testnet() {
-        run_testnet_test!(test_delete_key_no_key_left);
-    }
 }
