+++ tests/test_simple.rs
@@ -1,31 +1,21 @@
+///! Simply starts and runs TestNet for a while.
 #[cfg(test)]
 mod test {
     use std::sync::Arc;
     use std::thread;
     use std::time::Duration;
 
-    use network::proxy::predicate::FnProxyHandler;
+    use network::proxy::benchmark::BenchmarkHandler;
     use network::proxy::ProxyHandler;
     use primitives::transaction::TransactionBody;
     use testlib::node::{
         create_nodes, sample_two_nodes, Node, TEST_BLOCK_FETCH_LIMIT, TEST_BLOCK_MAX_SIZE,
     };
-    use testlib::test_helpers::wait;
+    use testlib::test_helpers::{heavy_test, wait};
 
     fn run_multiple_nodes(num_nodes: usize, num_trials: usize, test_prefix: &str, test_port: u16) {
         // Add proxy handlers to the pipeline.
-
-        // Custom handler example
-        let fn_proxy_handler = Arc::new(FnProxyHandler::new(|package| {
-            // Logic here...
-
-            // Return None to dismiss this channel,
-            // or Some(package) for passing to the next handler. Note that returned package don't need to be the
-            // same as the received package.
-            Some(package)
-        }));
-
-        let proxy_handlers: Vec<Arc<ProxyHandler>> = vec![fn_proxy_handler];
+        let proxy_handlers: Vec<Arc<ProxyHandler>> = vec![Arc::new(BenchmarkHandler::new())];
 
         let (init_balance, account_names, mut nodes) = create_nodes(
             num_nodes,
@@ -41,14 +31,13 @@ mod test {
             nodes[i].write().unwrap().start();
         }
 
-        thread::sleep(Duration::from_millis(1000));
-
         // Execute N trials. In each trial we submit a transaction to a random node i, that sends
         // 1 token to a random node j. We send transaction to node Then we wait for the balance change to propagate by checking
         // the balance of j on node k.
         let mut expected_balances = vec![init_balance; num_nodes];
-        let trial_duration = 10000;
-        for _ in 0..num_trials {
+        let trial_duration = 60000;
+        for trial in 0..num_trials {
+            println!("TRIAL #{}", trial);
             let (i, j) = sample_two_nodes(num_nodes);
             let (k, r) = sample_two_nodes(num_nodes);
             let nonce =
@@ -77,9 +66,13 @@ mod test {
         }
     }
 
-    /// Similar to `test_alphanet::test_4_10_multiple_nodes` to show custom proxy handlers usage.
     #[test]
-    fn test_custom_handler() {
-        run_multiple_nodes(3, 1, "custom_handler", 3200);
+    fn test_4_10_multiple_nodes() {
+        heavy_test(|| run_multiple_nodes(4, 10, "4_10", 3200));
+    }
+
+    #[test]
+    fn test_7_10_multiple_nodes() {
+        heavy_test(|| run_multiple_nodes(7, 10, "7_10", 3300));
     }
 }
